Update the Music on click
Made it so the music would start on-click, and stop on double-click.
Probably not the way we want it, but at least it’s progress.
I also made it stop the music on input change, and with the pause
button. There’s quite a bit of tweeking, but I think I’m on the right
track.